requirements.txt
pyyaml
click
coloredlogs
lxml
Jinja2
requests